Section 9
Just finished the first 2 lessons on this section and mostly learned about drivers and how to manage them. I also created a lab for the disk management section and created 2 drives and partitioned it with letters. Afterwards i created a raid 10 by creating 2 drives and using half of their space by making the striping raid first and then duplicating them with the mirrored tool in disk management creating a raid 10 array.